% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abimo_functions_am.R
\name{ABIMO_adapt_map}
\alias{ABIMO_adapt_map}
\title{adapt ABIMO output dbf-file to Berlin shape file}
\usage{
ABIMO_adapt_map(ABIMO_out, file_georef, out_file)
}
\arguments{
\item{ABIMO_out}{data.frame of ABIMO output file}

\item{file_georef}{path of dbf file that matches existing shape (incl. path)}

\item{out_file}{file path and file name for ordered ABIMO output file (to be linked to shape files)}
}
\value{
ordered dbf returned and written to out_file
}
\description{
changes order in dbf-file to match geographical
shape file of Berlin "Stadtstruktur"
}
