% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abimo_functions_am.R
\name{abimo_Berlin_average}
\alias{abimo_Berlin_average}
\title{calculate Berlin average of water balance terms}
\usage{
abimo_Berlin_average(abimo_df)
}
\arguments{
\item{abimo_df}{data.frame of ABIMO output file, merged with input file}
}
\value{
table with averages in mm of water balance components
}
\description{
multiplies each water balance component
by area, sums them up and divides the sum
by the total surface
}
