% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abimo_functions_am.R
\name{abimo_comb_in_out}
\alias{abimo_comb_in_out}
\title{join ABIMO in- and output files}
\usage{
abimo_comb_in_out(file_ABIMO_out, file_ABIMO_in)
}
\arguments{
\item{file_ABIMO_out}{path of ABIMO output file in dbf format (incl. path)}

\item{file_ABIMO_in}{path of ABIMO input file in dbf format (incl. path)}
}
\value{
data.frame with matched ABIMO in- and output data
}
\description{
join ABIMO in- and output files
}
