% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abimo_functions_am.R
\name{abimo_compare_output}
\alias{abimo_compare_output}
\title{Compares ABIMO-output-file to reference (or other ABIMO output file)}
\usage{
abimo_compare_output(x_reference, x_new)
}
\arguments{
\item{x_reference}{reference data frame with ABIMO output (can be ABIMO output or downloaded from Berlin geoportal)}

\item{x_new}{new ABIMO output to be compared to reference}
}
\value{
data.frame of column statistics; plots and evaluation open as pdf
}
\description{
Compares two ABIMO-output-files by
plotting parameters compared to 1:1 line into pdf-File and by
doing a simple column statistics.
}
