% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abimo_functions_am.R
\name{abimo_grwater_interflow}
\alias{abimo_grwater_interflow}
\title{calculate groundwater recharge and interflow}
\usage{
abimo_grwater_interflow(abimo_df)
}
\arguments{
\item{abimo_df}{data.frame of ABIMO output file, merged with input file}
}
\value{
input data.frame with two new columns "RI_K" and "INTERF"
}
\description{
uses correction factor to calculate groundwater recharge
from infiltration RI. Difference is interflow.
Requires a combined data.frame of ABIMO output and input,
e.g. by using function \code{\link{abimo_comb_in_out}}
}
