% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{abimo_input_2019}
\alias{abimo_input_2019}
\title{"Official" Abimo Input Data for Berlin, 2019, with streets}
\format{
A data frame with 25352 observations of 34variables:
\describe{
\item{CODE}{Variable 'CODE', see URL below}
\item{BEZIRK}{Variable 'BEZIRK', see URL below}
\item{STAGEB}{Variable 'STAGEB', see URL below}
\item{BLOCK}{Variable 'BLOCK', see URL below}
\item{TEILBLOCK}{Variable 'TEILBLOCK', see URL below}
\item{NUTZUNG}{Variable 'NUTZUNG', see URL below}
\item{TYP}{Variable 'TYP', see URL below}
\item{FLGES}{Variable 'FLGES', see URL below}
\item{VG}{Variable 'VG', see URL below}
\item{PROBAU}{Variable 'PROBAU', see URL below}
\item{PROVGU}{Variable 'PROVGU', see URL below}
\item{REGENJA}{Variable 'REGENJA', see URL below}
\item{REGENSO}{Variable 'REGENSO', see URL below}
\item{BELAG1}{Variable 'BELAG1', see URL below}
\item{BELAG2}{Variable 'BELAG2', see URL below}
\item{BELAG3}{Variable 'BELAG3', see URL below}
\item{BELAG4}{Variable 'BELAG4', see URL below}
\item{VGSTRASSE}{Variable 'VGSTRASSE', see URL below}
\item{STR_BELAG1}{Variable 'STR_BELAG1', see URL below}
\item{STR_BELAG2}{Variable 'STR_BELAG2', see URL below}
\item{STR_BELAG3}{Variable 'STR_BELAG3', see URL below}
\item{STR_BELAG4}{Variable 'STR_BELAG4', see URL below}
\item{KANAL}{Variable 'KANAL', see URL below}
\item{KANART}{Variable 'KANART', see URL below}
\item{KAN_BEB}{Variable 'KAN_BEB', see URL below}
\item{KAN_VGU}{Variable 'KAN_VGU', see URL below}
\item{KAN_STR}{Variable 'KAN_STR', see URL below}
\item{FELD_30}{Variable 'FELD_30', see URL below}
\item{FELD_150}{Variable 'FELD_150', see URL below}
\item{FLUR}{Variable 'FLUR', see URL below}
\item{STR_FLGES}{Variable 'STR_FLGES', see URL below}
\item{KOR_FL_N}{Variable 'KOR_FL_N', see URL below}
\item{AGEB1_NR}{Variable 'AGEB1_NR', see URL below}
\item{AGEB1}{Variable 'AGEB1', see URL below}
}
}
\source{
\url{https://www.berlin.de/umweltatlas/_assets/literatur/goedecke_et_al_abimo2019_doku.pdf}
}
\usage{
abimo_input_2019
}
\description{
The data was read from a file called "abimo_2019_mitstrassen.dbf"
using \code{\link[foreign]{read.dbf}}.
For a description of the input data, see the URL below.
}
\keyword{datasets}
