% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abimo_functions_am.R
\name{ABIMO_read_output}
\alias{ABIMO_read_output}
\title{Reads two ABIMO output files}
\usage{
ABIMO_read_output(SENSTADTUM_dbf, new_dbf)
}
\arguments{
\item{SENSTADTUM_dbf}{path of original SENSTADTUM-database}

\item{new_dbf}{path of new output-database}
}
\value{
output are two comparable (same dimensions and column names) data frames.
}
\description{
Reads a new ABIMO output file in dbase format. In addition the
original SENSTADTUM output file is read and made comparable. Alternatively
two new ABIMO output files can be read. Output are two comparable (same
dimensions and column names) data frames.
}
