% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abimo_functions_am.R
\name{abimo_xml_BER}
\alias{abimo_xml_BER}
\title{change "BERtoZero"-settings in Abimo config.xml to true}
\usage{
abimo_xml_BER(file_in = default_config(), file_out, line_BER = 56)
}
\arguments{
\item{file_in}{path and file name of abimo xml-input file, default is
\code{\link{default_config}}}

\item{file_out}{path and file name to write changed abimo xml-input file}

\item{line_BER}{line number in xml-file, where BERtoZero is defined, default is 56}
}
\value{
abimo xml-input file with changed BERtoZero-setting
}
\description{
as default irrigation of pervious areas
is assumed based on "Nutzung" and "Typ".
This function turns of irrigation for all areas
(BERtoZero = true)
}
