% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abimo_functions_am.R
\name{abimo_xml_evap}
\alias{abimo_xml_evap}
\title{change potential evaporation in Abimo config.xml}
\usage{
abimo_xml_evap(file_in = default_config(), file_out, evap_annual, evap_summer)
}
\arguments{
\item{file_in}{path and file name of abimo xml-input file, default is
\code{\link{default_config}}}

\item{file_out}{path and file name to write changed abimo xml-input file}

\item{evap_annual}{annual potential evaporation}

\item{evap_summer}{potential evaporation for summer months}
}
\value{
abimo xml-input file with changed potential evaporation
}
\description{
potential evaporation (annual and summer)
is a boundary condition defined in config.xml.
This function sets potential evaporation
to a given value for all surfaces (except lakes and rivers)
}
