% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abimo_functions_am.R
\name{add_ISU5_ID}
\alias{add_ISU5_ID}
\title{add ISU5 ID to dbf file from geoportal}
\usage{
add_ISU5_ID(
  x_no_ID,
  ID_dbf = "C:/Aendu_lokal/ABIMO_Paper/Daten/Karten/Basis_ISU5_Daten_2015/ISU5_ID.dbf"
)
}
\arguments{
\item{x_no_ID}{data.frame created from dbf file, downloaded from Berlin geoportal}

\item{ID_dbf}{path of dbf file including ids, default is set to local folder with ISU5-Ids}
}
\value{
data.frame of x_no_ID with a new column "ID"
}
\description{
the ID is an unambiguous identifier
for all blocks of the Berlin Geoportal
but is usually hidden
}
