% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abimo_functions_am.R
\name{appendSubToFile}
\alias{appendSubToFile}
\title{Add "SUB" field to dbf-File}
\usage{
appendSubToFile(filename)
}
\arguments{
\item{filename}{Path of file name of data.frame}
}
\value{
dbf file with sub field
}
\description{
Adds "SUB" field to the end of an existing  file, as expected by some
older applications (such as input-dbf-file for ABIMO)
function by grandmaster HAUKESON
}
