% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_bagrov_curve.R
\name{calculate_bagrov_curve}
\alias{calculate_bagrov_curve}
\title{Calculate the curve(s) of the Bagrov relation}
\usage{
calculate_bagrov_curve(
  effectivity,
  P_over_Ep_max = 4,
  Ep = 650,
  delta_Ea = 0.1,
  dbg = FALSE
)
}
\arguments{
\item{effectivity}{Bagrov effectivity value (= n-value)}

\item{P_over_Ep_max}{maximum value of the ratio P/Ep with precipitation P and
potential evaporation Ep.}

\item{Ep}{potential evaporation in mm/a (default: 650)}

\item{delta_Ea}{step of actual evaporation (default: 0.1)}

\item{dbg}{logical indicating whether or not to show debug messages}
}
\value{
data frame with columns \code{P_over_Ep} (P/Ep), \code{Ea_over_Ep}
(Ea/Ep), \code{effectivity} (n-value)
}
\description{
Calculate the curve(s) of the Bagrov relation
}
\examples{
\dontrun{
bagrov_data <- calculate_bagrov_curve(1:2)
plot(bagrov_data$P_over_Ep, bagrov_data$Ea_over_Ep)
}
}
