% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure.R
\name{create_configurator}
\alias{create_configurator}
\title{Create Configuration Object from Abimo Configuration File}
\usage{
create_configurator(xml_file = NULL)
}
\arguments{
\item{xml_file}{path to "config.xml". By default a configuration file that is
stored in this package (\code{system.file("extdata/config.xml")}) is used.}
}
\value{
object of class "abimoConfig"
}
\description{
Create Configuration Object from Abimo Configuration File
}
