% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_bagrov_curves_from_abimo}
\alias{get_bagrov_curves_from_abimo}
\title{Get Bagrov curves from Abimo software}
\usage{
get_bagrov_curves_from_abimo()
}
\value{
data frame with columns \code{P_over_Ep} (P/Ep), \code{Ea_over_Ep}
(Ea/Ep), \code{effectivity} (n-value)
}
\description{
Call Abimo with the --write-bagrov-table argument being set and convert the
console output to a data frame
}
