% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure.R
\name{get_xpaths}
\alias{get_xpaths}
\title{Return List of XPath Expressions to Address Config Elements}
\usage{
get_xpaths()
}
\value{
List structure containing XPath expressions (see e.g.
\url{https://www.w3schools.com/xml/xpath_intro.asp})
}
\description{
Return List of XPath Expressions to Address Config Elements
}
