% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_abimo.R
\name{install_abimo}
\alias{install_abimo}
\title{Install ABIMO}
\usage{
install_abimo(
  tag = latest_abimo_version(),
  arch = get_architecture_suffix(),
  ...
)
}
\arguments{
\item{tag}{tag of ABIMO version to be installed, e.g. "v3.2.2"}

\item{arch}{target system architecture, one of "windows", "linux", "macos"}

\item{\dots}{further arguments passed to \code{kwb.abimo:::download_assets}}
}
\description{
Install ABIMO
}
