% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_abimo_intermediate_results_from_log.R
\name{read_abimo_intermediate_results_from_log}
\alias{read_abimo_intermediate_results_from_log}
\title{Read Intermediate Results from Log File}
\usage{
read_abimo_intermediate_results_from_log(
  file = file.path(tempdir(), "abimo_input_result.log")
)
}
\arguments{
\item{file}{path to log file}
}
\description{
Read Intermediate Results from Log File
}
