% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_2.R
\name{read_config}
\alias{read_config}
\title{Read Abimo Configuration from XML File}
\usage{
read_config(file = default_config())
}
\arguments{
\item{file}{path to XML file}
}
\description{
Read Abimo Configuration from XML File
}
