% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_intermediate_results_from_log.R
\name{read_intermediate_results_from_log}
\alias{read_intermediate_results_from_log}
\title{Read Intermediate Results from Log File}
\usage{
read_intermediate_results_from_log(
  file = file.path(tempdir(), "abimo_input_result.log"),
  pattern_remove = "Start|unknown|angenommen|nicht definiert|std::"
)
}
\arguments{
\item{file}{path to log file}

\item{pattern_remove}{regular expression matching lines to remove from the
log file before looking for "variable=expression" assignments}
}
\description{
Read Intermediate Results from Log File
}
