% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_abimo.R
\name{run_abimo}
\alias{run_abimo}
\title{Run Abimo with Input Data or Input File}
\usage{
run_abimo(
  input_file = NULL,
  input_data = NULL,
  output_file = NULL,
  config_file = NULL,
  config = NULL,
  tag = latest_abimo_version(),
  read_intermediates = FALSE
)
}
\arguments{
\item{input_file}{path to input dbf file}

\item{input_data}{data frame from which a temporary input file is to be
generated}

\item{output_file}{path to output file. By default the output file has the
same name as the input file with "_result" appended}

\item{config_file}{optional. Path to config.xml file}

\item{config}{optional. Configuration object of class "abimoConfig", as
returned by \code{create_configurator}. If given, \code{config_file} is
ignored.}

\item{tag}{version tag of Abimo release to be used, see
\url{https://github.com/KWB-R/abimo/releases}.}

\item{read_intermediates}{if \code{TRUE} the values of intermediate variables
are read from the log file (if applicable). The default is \code{FALSE}.}
}
\value{
data frame, read from dbf file that was created by Abimo.exe.
If \code{read_intermediates} is \code{TRUE}, intermediate results are
returned in the attribute "intermediates"
}
\description{
Run Abimo with Input Data or Input File
}
