% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_abimo.R
\name{run_abimo_command_line}
\alias{run_abimo_command_line}
\title{Run Abimo on the Command Line}
\usage{
run_abimo_command_line(args, tag = "v3.3.0")
}
\arguments{
\item{args}{vector of arguments to be passed to Abimo}

\item{tag}{version tag of Abimo release to be used, see
\url{https://github.com/KWB-R/abimo/releases}}
}
\value{
The function returns what Abimo.exe sent to the standard output (as a
vector of character).
}
\description{
Run Abimo on the Command Line
}
