% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abimo_functions_am.R
\name{write.dbf.abimo}
\alias{write.dbf.abimo}
\title{writes data.frame into ABIMO-dbf}
\usage{
write.dbf.abimo(df_name, new_dbf)
}
\arguments{
\item{df_name}{name of data.frame}

\item{new_dbf}{path of new ABIMO-input file to be written (.dbf)}
}
\value{
dbf file that can be processed by ABIMO
}
\description{
Saves an existing data.frame into dBase-format
and adds "SUB" field to the end of the file
as required by ABIMO
}
